# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Чат для Системы HRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/chats": {
            "get": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Возвращает все чаты, участником которых является текущий пользователь. Для личных (direct) чатов имя автоматически устанавливается как имя собеседника.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "чаты"
                ],
                "summary": "Получить чаты пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Chat"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Создает личный (direct) или групповой (group) чат. Личный чат требует ровно 1 собеседника (нельзя с самим собой). Групповой чат требует минимум 3 участников.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "чаты"
                ],
                "summary": "Создать новый чат",
                "parameters": [
                    {
                        "description": "Параметры создания чата",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateChatDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Chat"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
        "/chats/{id}/avatar": {
            "get": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Возвращает изображение аватарки указанного чата",
                "tags": [
                    "чаты"
                ],
                "summary": "Получить аватар чата по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Загружает изображение аватарки группового или личного чата",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "чаты"
                ],
                "summary": "Загрузить аватар чата",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения аватарки",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Chat"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/chats/{id}/files": {
            "post": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Загружает файл и отправляет сообщение с прикрепленным файлом в указанный чат",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "чаты"
                ],
                "summary": "Загрузить файл в чат",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID чата",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл для загрузки",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Необязательная подпись к файлу",
                        "name": "caption",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Message"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Chat": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_global": {
                    "type": "boolean"
                },
                "last_message": {
                    "$ref": "#/definitions/domain.Message"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ChatMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "description": "\"direct\" or \"group\"",
                    "type": "string"
                },
                "unread_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ChatMember": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                },
                "last_read_message_id": {
                    "type": "integer"
                },
                "role": {
                    "description": "\"owner\", \"admin\", \"member\"",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateChatDTO": {
            "type": "object",
            "properties": {
                "member_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Project Team"
                },
                "type": {
                    "type": "string",
                    "example": "group"
                }
            }
        },
...
        "http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "LoginHeaderAuth": {
            "description": "Передайте имя пользователя или email для авторизации (например: amin, fed, ernest или far@gmail.com)",
            "type": "apiKey",
            "name": "Login",
            "in": "header"
        }
    }
}